+++ examples/post_with_image.rs
@@ -8,6 +8,7 @@ use jacquard::oauth::loopback::LoopbackConfig;
 use jacquard::oauth::types::AuthorizeOptions;
 use jacquard::types::blob::MimeType;
 use jacquard::types::string::Datetime;
+use jacquard_oauth::scopes::Scopes;
 use miette::IntoDiagnostic;
 use smol_str::SmolStr;
 use std::path::PathBuf;
@@ -44,7 +45,14 @@ async fn main() -> miette::Result<()> {
     let Some(session) = oauth
         .resume_or_login_with_local_server(
             &hint,
-            AuthorizeOptions::default(),
+            AuthorizeOptions::default().with_scopes(
+                Scopes::builder()
+                    .include_aud(
+                        "app.bsky.authCreatePosts",
+                        "did:web:public.api.bsky.app#bsky_appview",
+                    )?
+                    .build()?,
+            ),
             LoopbackConfig::default(),
         )
         .await?