+++ examples/update_profile.rs
@@ -6,6 +6,7 @@ use jacquard::oauth::client::OAuthClient;
 use jacquard::oauth::loopback::LoopbackConfig;
 use jacquard::oauth::types::AuthorizeOptions;
 use jacquard::types::string::AtUri;
+use jacquard_oauth::scopes::Scopes;
 use smol_str::SmolStr;
 
 #[derive(Parser, Debug)]
@@ -36,7 +37,14 @@ async fn main() -> miette::Result<()> {
     let Some(session) = oauth
         .resume_or_login_with_local_server(
             &hint,
-            AuthorizeOptions::default(),
+            AuthorizeOptions::default().with_scopes(
+                Scopes::builder()
+                    .include_aud(
+                        "app.bsky.authManageProfile",
+                        "did:web:public.api.bsky.app#bsky_appview",
+                    )?
+                    .build()?,
+            ),
             LoopbackConfig::default(),
         )
         .await?